Numpy_Basics
NumPy is a powerful Python library for numerical computing. It provides a multidimensional array object, along with a collection of functions and tools for working with arrays. NumPy enables efficient and high-performance operations on large datasets and arrays, including mathematical, logical, shape manipulation, and statistical operations. It is widely used in scientific and data-intensive applications for tasks such as data analysis, modeling, simulation, and machine learning. With NumPy, users can perform complex computations and manipulations on arrays with ease